requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow               # image decode
torch==2.2.2+cpu     # CPU wheel (~160 MB) – enough for SigLIP
--extra-index-url https://download.pytorch.org/whl/cpu

transformers==4.41.1
accelerate==0.30.1
